Move Model classes into there own package under core
