+++ src/main/java/com/karlnosworthy/poijoi/io/xls/XLSSpreadsheetReader.java
@@ -16,11 +16,11 @@ import org.apache.poi.hssf.usermodel.HSSFWorkbook;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.DataFormatter;
 
-import com.karlnosworthy.poijoi.io.ColumnDefinition;
-import com.karlnosworthy.poijoi.io.ColumnDefinition.ColumnType;
-import com.karlnosworthy.poijoi.io.PoijoiMetaData;
+import com.karlnosworthy.poijoi.core.model.ColumnDefinition;
+import com.karlnosworthy.poijoi.core.model.PoijoiMetaData;
+import com.karlnosworthy.poijoi.core.model.TableDefinition;
+import com.karlnosworthy.poijoi.core.model.ColumnDefinition.ColumnType;
 import com.karlnosworthy.poijoi.io.Reader;
-import com.karlnosworthy.poijoi.io.TableDefinition;
 
 public final class XLSSpreadsheetReader implements Reader {
